Scanning Tools/requirements.txt
# Requirements file for Heritage Digital Twinning - Scanning Tools
# Install with: pip install -r requirements.txt

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Computer vision and image processing
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
Pillow>=8.3.0

# 3D point cloud processing
open3d>=0.13.0

# Machine learning and clustering
scikit-learn>=1.0.0

# Data handling and serialization
pandas>=1.3.0
json5>=0.9.0

# Geospatial processing
geopy>=2.2.0
folium>=0.12.0
pyproj>=3.2.0

# Image metadata and EXIF
exifread>=2.3.2
pillow-heif>=0.3.0

# Photogrammetry integration
pycolmap>=0.1.0  # Optional: for COLMAP integration

# Mesh processing
trimesh>=3.9.0
pymeshlab>=2021.10  # Optional: for advanced mesh processing

# Progress bars and logging
tqdm>=4.62.0
coloredlogs>=15.0

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0